LayoutTests/fast/xmlhttprequest/xmlhttprequest-access-self-as-file.html

<!DOCTYPE html>
<html>
<head>
<script src="../../resources/js-test.js"></script>
<script>
description("Tests that you cannot XHR to the current file URL.");

if (window.testRunner) {
    testRunner.dumpAsText();
    testRunner.dumpChildFramesAsText();
    testRunner.setAllowUniversalAccessFromFileURLs(false);
}
</script>
</head>
<body>
<iframe src="resources/xmlhttprequest-access-self-as-file-real.html"></iframe>
</body>
</html>
